move TestListS3Paths class from test_utils to new test_s3 testing module
